requirements.txt
click==6.7
Flask==0.12
Flask-SQLAlchemy==2.1
Flask-Migrate==2.0.3
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
Werkzeug==0.11.15
line-bot-sdk==1.5.0
beautifulsoup4==4.5.3
requests==2.12.4
